Modifying the sample sentence to be consistent
The sentence "Probabilistic graphical models are fun. They are also powerful." contains 9 words not 10. 
The word "are" is repeated twice, so you cannot say "each word appearing once, and hence, each word in the corpus is assigned a probability of 0.1."
My suggestion is: "Probabilistic graphical models are fun. They can also be powerful."